refactor: 저장 및 댓글 수정
게시글 로드시 접속자 좋아요 및 저장 여부 표시
댓글 좋아요 여부 추가
저장 목록 불러오기
userId -> Authentication으로 수정

Resolves:
Related to: S11P12D208-111

+++ backend/src/main/java/com/campforest/backend/board/controller/BoardController.java
@@ -29,6 +29,7 @@ import com.campforest.backend.common.ApiResponse;
 import com.campforest.backend.common.ErrorCode;
 import com.campforest.backend.config.s3.S3Service;
 import com.campforest.backend.product.dto.ProductRegistDto;
+import com.campforest.backend.product.dto.ProductUpdateDto;
 import com.campforest.backend.user.model.Users;
 import com.campforest.backend.user.service.UserService;
 
@@ -97,9 +98,7 @@ public class BoardController {
 		try {
 			if (page < 0 || size <= 0) {
 				return ApiResponse.createError(ErrorCode.INVALID_PAGE_NUMBER);
-			}
-			else if (authentication == null) {
-				System.out.println("유저 없을때");
+			} else if (authentication == null) {
 				Page<BoardResponseDto> boardResponseDtos = boardService.getAllBoards(-1L, page, size);
 
 				return ApiResponse.createSuccess(boardResponseDtos, "비로그인 게시글 목록 조회 성공하였습니다");
@@ -113,7 +112,6 @@ public class BoardController {
 				return ApiResponse.createSuccess(boardResponseDtos, "게시글 목록 조회 성공하였습니다");
 			}
 		} catch (Exception e) {
-			System.out.println(e.getMessage());
 			return ApiResponse.createError(ErrorCode.INTERNAL_SERVER_ERROR);
 		}
 	}
@@ -130,20 +128,18 @@ public class BoardController {
 		try {
 			if (page < 0 || size <= 0) {
 				return ApiResponse.createError(ErrorCode.INVALID_PAGE_NUMBER);
-			}
-			else if (authentication == null) {
-				Page<BoardResponseDto> boardResponseDtos = boardService.getUserBoards(-1L,userId, page, size);
+			} else if (authentication == null) {
+				Page<BoardResponseDto> boardResponseDtos = boardService.getUserBoards(-1L, userId, page, size);
 				return ApiResponse.createSuccess(boardResponseDtos, "비로그인 게시글 사용자별 조회 성공하였습니다");
+			} else {
+				Users user = userService.findByEmail(authentication.getName())
+					.orElseThrow(() -> new Exception("유저 정보 조회 실패"));
+				;
+				Long nowId = user.getUserId();
+				Page<BoardResponseDto> boardResponseDtos = boardService.getUserBoards(nowId, userId, page, size);
+				return ApiResponse.createSuccess(boardResponseDtos, "게시글 사용자별 조회에 성공하였습니다");
 			}
-			else{
-			Users user = userService.findByEmail(authentication.getName())
-				.orElseThrow(() -> new Exception("유저 정보 조회 실패"));;
-			Long nowId= user.getUserId();
-			Page<BoardResponseDto> boardResponseDtos = boardService.getUserBoards(nowId, userId, page, size);
-			return ApiResponse.createSuccess(boardResponseDtos, "게시글 사용자별 조회에 성공하였습니다");
-		}
-		}
-			catch (Exception e) {
+		} catch (Exception e) {
 			return ApiResponse.createError(ErrorCode.USER_NOT_FOUND);
 		}
 	}
@@ -158,19 +154,17 @@ public class BoardController {
 		try {
 			if (page < 0 || size <= 0) {
 				return ApiResponse.createError(ErrorCode.INVALID_PAGE_NUMBER);
-			}
-			else if (authentication == null) {
+			} else if (authentication == null) {
 				Page<BoardResponseDto> boardResponseDtos = boardService.getTitleBoards(-1L, title, page, size);
 				return ApiResponse.createSuccess(boardResponseDtos, "비로그인 게시글 제목으로 검색에 성공하였습니다");
-			}
-			else {
+			} else {
 				Users user = userService.findByEmail(authentication.getName())
 					.orElseThrow(() -> new Exception("유저 정보 조회 실패"));
-				Long nowId= user.getUserId();
+				Long nowId = user.getUserId();
 				Page<BoardResponseDto> boardResponseDtos = boardService.getTitleBoards(nowId, title, page, size);
 				return ApiResponse.createSuccess(boardResponseDtos, "게시글 제목으로 검색에 성공하였습니다");
 			}
-			} catch (Exception e) {
+		} catch (Exception e) {
 			return ApiResponse.createError(ErrorCode.INVALID_BOARD_CATEGORY);
 		}
 	}
@@ -185,28 +179,72 @@ public class BoardController {
 		try {
 			if (page < 0 || size <= 0) {
 				return ApiResponse.createError(ErrorCode.INVALID_PAGE_NUMBER);
-			}
-			else if (authentication == null) {
+			} else if (authentication == null) {
 				Page<BoardResponseDto> boardResponseDtos = boardService.getCategoryBoards(-1L, category, page, size);
 				return ApiResponse.createSuccess(boardResponseDtos, "비로그인 게시글 카테고리별 조회에 성공하였습니다");
 
-			}
-			else {
+			} else {
 				Users user = userService.findByEmail(authentication.getName())
-					.orElseThrow(() -> new Exception("유저 정보 조회 실패"));;
-				Long nowId= user.getUserId();
+					.orElseThrow(() -> new Exception("유저 정보 조회 실패"));
+				;
+				Long nowId = user.getUserId();
 				Page<BoardResponseDto> boardResponseDtos = boardService.getCategoryBoards(nowId, category, page, size);
 				return ApiResponse.createSuccess(boardResponseDtos, "게시글 카테고리별 조회에 성공하였습니다");
 			}
-			} catch (Exception e) {
+		} catch (Exception e) {
+			return ApiResponse.createError(ErrorCode.INVALID_BOARD_CATEGORY);
+		}
+	}
+
+	//저장 목록 불러오기
+	@GetMapping("/savedboard")
+	public ApiResponse<?> getSavedBoard(
+		Authentication authentication,
+		@RequestParam(defaultValue = "0") int page,
+		@RequestParam(defaultValue = "10") int size) {
+		try {
+			if (page < 0 || size <= 0) {
+				return ApiResponse.createError(ErrorCode.INVALID_PAGE_NUMBER);
+			}
+			if (authentication == null) {
+				return ApiResponse.createError(ErrorCode.INVALID_AUTHORIZED);
+			} else {
+				Users user = userService.findByEmail(authentication.getName())
+					.orElseThrow(() -> new Exception("유저 정보 조회 실패"));
+				;
+				Long nowId = user.getUserId();
+				Page<BoardResponseDto> boardResponseDtos = boardService.getSavedBoards(nowId, page, size);
+				return ApiResponse.createSuccess(boardResponseDtos, "저장한 게시 조회에 성공하였습니다");
+			}
+		} catch (Exception e) {
 			return ApiResponse.createError(ErrorCode.INVALID_BOARD_CATEGORY);
 		}
 	}
 
 	//게시물 수정
 	@PutMapping
-	public ApiResponse<?> modifyBoard(@RequestParam Long boardId, @RequestBody BoardRequestDto boardRequestDto) {
+	public ApiResponse<?> modifyBoard(
+		Authentication authentication,
+		@RequestParam Long boardId,
+		@RequestPart(value = "files", required = false) MultipartFile[] files,
+		@RequestPart(value = "boardRequestDto") BoardRequestDto boardRequestDto) {
 		try {
+			Users user = userService.findByEmail(authentication.getName())
+				.orElseThrow(() -> new Exception("유저 정보 조회 실패"));
+			if (!boardRequestDto.getUserId().equals(user.getUserId())) {
+				return ApiResponse.createError(ErrorCode.INVALID_AUTHORIZED);
+			}
+			List<String> imageUrls = new ArrayList<>();
+			if (files != null) {
+				for (MultipartFile file : files) {
+					String extension = file.getOriginalFilename()
+						.substring(file.getOriginalFilename().lastIndexOf("."));
+					String fileUrl = s3Service.upload(file.getOriginalFilename(), file, extension);
+					imageUrls.add(fileUrl);
+				}
+			}
+			boardRequestDto.setUserId(user.getUserId());
+			boardRequestDto.setImageUrls(imageUrls);
 			boardService.modifyBoard(boardId, boardRequestDto);
 			return ApiResponse.createSuccessWithNoContent("게시물 수정에 성공하였습니다.");
 		} catch (Exception e) {
@@ -216,8 +254,14 @@ public class BoardController {
 
 	//게시글 삭제
 	@DeleteMapping
-	public ApiResponse<?> deleteBoard(@RequestParam Long boardId) {
+	public ApiResponse<?> deleteBoard(Authentication authentication, @RequestParam Long boardId) {
 		try {
+			Users user = userService.findByEmail(authentication.getName())
+				.orElseThrow(() -> new Exception("유저 정보 조회 실패"));
+			if (!boardService.getBoard(boardId).getUserId().equals(user.getUserId())) {
+				return ApiResponse.createError(ErrorCode.INVALID_AUTHORIZED);
+			}
+
 			boardService.deleteBoard(boardId);
 			return ApiResponse.createSuccessWithNoContent("게시글 삭제 성공하였습니다");
 		} catch (Exception e) {
@@ -227,8 +271,11 @@ public class BoardController {
 
 	//게시글 좋아요
 	@PostMapping("/like")
-	public ApiResponse<?> likeBoard(@RequestParam Long boardId, @RequestParam Long userId) {
+	public ApiResponse<?> likeBoard(Authentication authentication, @RequestParam Long boardId) {
 		try {
+			Users user = userService.findByEmail(authentication.getName())
+				.orElseThrow(() -> new Exception("유저 정보 조회 실패"));
+			Long userId = user.getUserId();
 			if (boardService.checkLike(boardId, userId)) {
 				return ApiResponse.createError(ErrorCode.LIKE_ALREADY_EXISTS);
 			}
@@ -241,14 +288,20 @@ public class BoardController {
 
 	//게시글 좋아요 삭제
 	@DeleteMapping("like")
-	public ApiResponse<?> deleteLike(@RequestParam Long boardId, @RequestParam Long userId) {
+	public ApiResponse<?> deleteLike(Authentication authentication, @RequestParam Long boardId) {
+		try {
+			Users user = userService.findByEmail(authentication.getName())
+				.orElseThrow(() -> new Exception("유저 정보 조회 실패"));
 
-		if (!boardService.checkLike(boardId, userId)) {
-			return ApiResponse.createError(ErrorCode.LIKE_NOT_FOUND);
+			Long userId = user.getUserId();
+			if (!boardService.checkLike(boardId, userId)) {
+				return ApiResponse.createError(ErrorCode.LIKE_NOT_FOUND);
+			}
+			boardService.deleteLike(boardId, userId);
+			return ApiResponse.createSuccess(boardService.getBoard(boardId).getLikeCount(), "게시글 좋아요 삭제 성공하였습니다");
+		} catch (Exception e) {
+			return ApiResponse.createError(ErrorCode.INTERNAL_SERVER_ERROR);
 		}
-		boardService.deleteLike(boardId, userId);
-
-		return ApiResponse.createSuccess(boardService.getBoard(boardId).getLikeCount(), "게시글 좋아요 삭제 성공하였습니다");
 
 	}
 
@@ -265,8 +318,12 @@ public class BoardController {
 
 	//게시글 저장
 	@PostMapping("/save")
-	public ApiResponse<?> saveBoard(@RequestParam Long boardId, @RequestParam Long userId) {
+	public ApiResponse<?> saveBoard(Authentication authentication, @RequestParam Long boardId) {
 		try {
+			Users user = userService.findByEmail(authentication.getName())
+				.orElseThrow(() -> new Exception("유저 정보 조회 실패"));
+
+			Long userId = user.getUserId();
 			if (boardService.checkSave(boardId, userId)) {
 				return ApiResponse.createError(ErrorCode.SAVE_ALREADY_EXISTS);
 			}
@@ -278,8 +335,12 @@ public class BoardController {
 	}
 
 	@DeleteMapping("/save")
-	public ApiResponse<?> deleteSave(@RequestParam Long boardId, @RequestParam Long userId) {
+	public ApiResponse<?> deleteSave(Authentication authentication, @RequestParam Long boardId) {
 		try {
+			Users user = userService.findByEmail(authentication.getName())
+				.orElseThrow(() -> new Exception("유저 정보 조회 실패"));
+
+			Long userId = user.getUserId();
 			boardService.deleteSave(boardId, userId);
 			return ApiResponse.createSuccessWithNoContent("저장 삭제 성공하였습니다");
 		} catch (Exception e) {
@@ -289,8 +350,14 @@ public class BoardController {
 
 	//게시글에 댓글 작성
 	@PostMapping("/comment")
-	public ApiResponse<?> writeComment(@RequestParam Long boardId, @RequestBody CommentRequestDto commentRequestDto) {
+	public ApiResponse<?> writeComment(
+		Authentication authentication,
+		@RequestParam Long boardId,
+		@RequestBody CommentRequestDto commentRequestDto) {
 		try {
+			Users users = userService.findByEmail(authentication.getName())
+				.orElseThrow(() -> new Exception("유저 정보 조회 실패"));
+
 			boardService.writeComment(boardId, commentRequestDto);
 			return ApiResponse.createSuccessWithNoContent("댓글 작성 성공");
 		} catch (Exception e) {
@@ -300,10 +367,25 @@ public class BoardController {
 
 	//게시글에 달려있는 댓글 목록 조회
 	@GetMapping("/comment")
-	public ApiResponse<?> getComment(@RequestParam Long boardId) {
+	public ApiResponse<?> getComment(
+		Authentication authentication,
+		@RequestParam Long boardId,
+		@RequestParam(defaultValue = "0") int page,
+		@RequestParam(defaultValue = "10") int size) {
 		try {
-			List<CommentResponseDto> commentResponseDtos = boardService.getComment(boardId);
-			return ApiResponse.createSuccess(commentResponseDtos, "댓글 게시글별 조회에 성공하였습니다");
+			if (page < 0 || size <= 0) {
+				return ApiResponse.createError(ErrorCode.INVALID_PAGE_NUMBER);
+			} else if (authentication == null) {
+				Page<CommentResponseDto> commentResponseDtos = boardService.getComments(-1L, boardId, page, size);
+				return ApiResponse.createSuccess(commentResponseDtos, "비로그인 댓글 게시글별 조회 성공하였습니다");
+			} else {
+				Users user = userService.findByEmail(authentication.getName())
+					.orElseThrow(() -> new Exception("유저 정보 조회 실패"));
+				;
+				Long nowId = user.getUserId();
+				Page<CommentResponseDto> commentResponseDtos = boardService.getComments(nowId, boardId, page, size);
+				return ApiResponse.createSuccess(commentResponseDtos, "댓글 게시글별 조회에 성공하였습니다");
+			}
 		} catch (Exception e) {
 			return ApiResponse.createError(ErrorCode.BOARD_NOT_FOUND);
 		}
@@ -322,8 +404,15 @@ public class BoardController {
 
 	//댓글 삭제
 	@DeleteMapping("/comment")
-	public ApiResponse<?> deleteComment(@RequestParam Long commentId) {
+	public ApiResponse<?> deleteComment(
+		Authentication authentication,
+		@RequestParam Long commentId) {
 		try {
+			Users user = userService.findByEmail(authentication.getName())
+				.orElseThrow(() -> new Exception("유저 정보 조회 실패"));
+			if (!boardService.getCommentById(commentId).getCommentWriterId().equals(user.getUserId())) {
+				return ApiResponse.createError(ErrorCode.INVALID_AUTHORIZED);
+			}
 			boardService.deleteComment(commentId);
 			return ApiResponse.createSuccessWithNoContent("댓글 삭제 성공하였습니다");
 		} catch (Exception e) {
@@ -344,13 +433,17 @@ public class BoardController {
 
 	//댓글 좋아요
 	@PostMapping("/commentlike")
-	public ApiResponse<?> likeComment(@RequestParam Long commentId, @RequestParam Long userId) {
+	public ApiResponse<?> likeComment(Authentication authentication, @RequestParam Long commentId) {
 		try {
+			Users user = userService.findByEmail(authentication.getName())
+				.orElseThrow(() -> new Exception("유저 정보 조회 실패"));
+
+			Long userId = user.getUserId();
 			if (boardService.checkCommentLike(commentId, userId)) {
 				return ApiResponse.createError(ErrorCode.COMMENT_LIKE_ALREADY_EXISTS);
 			} else {
 				boardService.likeComment(commentId, userId);
-				return ApiResponse.createSuccessWithNoContent("댓글 좋아요 성공하였습니다");
+				return ApiResponse.createSuccess(boardService.countCommentLike(commentId), "댓글 좋아요 성공하였습니다");
 			}
 		} catch (Exception e) {
 			return ApiResponse.createError(ErrorCode.INTERNAL_SERVER_ERROR);
@@ -360,10 +453,16 @@ public class BoardController {
 
 	//댓글 좋아요 삭제
 	@DeleteMapping("/commentlike")
-	public ApiResponse<?> deleteCommentLike(@RequestParam Long commentId, @RequestParam Long userId) {
+	public ApiResponse<?> deleteCommentLike(Authentication authentication, @RequestParam Long commentId) {
 		try {
+			Users user = userService.findByEmail(authentication.getName())
+				.orElseThrow(() -> new Exception("유저 정보 조회 실패"));
+
+			Long userId = user.getUserId();
+
 			boardService.deleteCommentLike(commentId, userId);
-			return ApiResponse.createSuccessWithNoContent("댓글 좋아요 삭제 성공하였습니다");
+			return ApiResponse.createSuccess(boardService.countCommentLike(commentId), "댓글 삭제 성공하였습니다");
+
 		} catch (Exception e) {
 			return ApiResponse.createError(ErrorCode.COMMENT_LIKE_NOT_FOUND);
 		}
